requirements.txt
Cython
attrs==22.2.0
ConfigParser==5.3.0
contextlib2==21.6.0
cryptography==40.0.2
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
ipython==8.13.1
ipywidgets==8.0.6
Jinja2==3.1.2
keyring==23.13.1
langchain==0.0.153
lockfile==0.12.2
mock==5.0.2
openai==0.27.2
Pillow==9.5.0
protobuf==3.20.3
pyOpenSSL==23.1.1
python-dotenv==1.0.0
railroad==0.5.0
Sphinx==7.0.0
tornado==6.3.1
trove_classifiers==2023.4.29
truststore==0.7.0
urllib3_secure_extra==0.1.0
chromadb==0.3.22
streamlit

